pyproject.toml
[project]
name = "cohort5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langfuse>=3.8.1",
    "pillow>=12.1.1",
    "pydantic-ai-slim[logfire,openai]>=1.9.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
